HUGE amount of changes -- read full text.
Turned off spellcheck in PyCharm's settings (was really bothering me).

Renamed a bunch of stuff and moved certain tasks into functions. Also added error checking to the reddit login and the SMTP connection.

Got rid of all the globals (except for a global prefs variable, which makes my life a little easier) and just passed the reddit and smtp connection instances to the functions that need them.

Prefs are now a dictionary instead of separate variables. Also added a sleep_time property to config.ini to control how often the bot checks for new posts.

If the bot gets disconnected somehow during operation, it'll try to reconnect instead of crashing.