v1.1 - Improved error-handling + lots of small changes.
Updated version to v1.1.

If an error occurs when sending an email, it is now caught and alerted to the user properly.

In get_smtp_connection(), smtp_connection was never quit and the sys.exit() call was repeated every time an exception was caught. To fix this, these behaviours have been moved into a local function called handle_error().

When a giveaway is detected, the link to the giveaway is printed to the console. There is also now a notification when an email alert has been successfully sent.

Removed the 'Checking {post title}' console alert.